+++ spec/spec_helper.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require 'rspec_junit_formatter'
 require 'process_settings'
 
 require 'pry'
@@ -7,6 +8,9 @@ require 'coveralls'
 Coveralls.wear!
 
 RSpec.configure do |config|
+  config.add_formatter  :progress
+  config.add_formatter  RspecJunitFormatter, ENV['JUNIT_OUTPUT'] || 'spec/reports/rspec.xml'
+
   config.expect_with :rspec do |expectations|
     expectations.include_chain_clauses_in_custom_matcher_descriptions = true
   end